Added input function and other small changes
Created an input function called inputColor that has the same functionality as printColor but used to add color to input text. Also changed all "2's" into "light" to be more specific that
the alternative colors are lighter. Lastly, changed "BEIGE" to "CYAN" as it was accidentally given the wrong label.